+++ lib/logstash-logger/device/connectable.rb
@@ -5,6 +5,8 @@ module LogStashLogger
     class Connectable < Base
       include LogStashLogger::Buffer
 
+      attr_accessor :buffer_logger
+
       def initialize(opts = {})
         super
 
@@ -40,9 +42,12 @@ module LogStashLogger
             true
           end
 
+        @buffer_logger = opts[:buffer_logger]
+
         buffer_initialize(
           max_items: @buffer_max_items,
           max_interval: @buffer_max_interval,
+          logger: buffer_logger,
           autoflush: @sync,
           drop_messages_on_flush_error: @drop_messages_on_flush_error,
           drop_messages_on_full_buffer: @drop_messages_on_full_buffer,
@@ -51,7 +56,7 @@ module LogStashLogger
       end
 
       def write(message)
-        buffer_receive message, @buffer_group
+        buffer_receive(message, @buffer_group) unless message.nil?
       end
 
       def flush(*args)
@@ -100,6 +105,11 @@ module LogStashLogger
       # Implemented by subclasses
       def connect
         fail NotImplementedError
+      end
+
+      def reset
+        reset_buffer
+        close(flush: false)
       end
 
       def reconnect