+++ lib/logstash-logger/formatter/json_lines.rb
@@ -1,9 +1,15 @@
 module LogStashLogger
   module Formatter
     class JsonLines < Base
-      def call(severity, time, progname, message)
-        super
-        "#{@event.to_json}\n"
+      private
+
+      def format_event(event)
+        "#{event.to_json}\n"
+      rescue Encoding::UndefinedConversionError,
+             Encoding::InvalidByteSequenceError,
+             JSON::GeneratorError => e
+        log_error(e)
+        "#{force_utf8_encoding(event).to_json}\n"
       end
     end
   end