+++ spec/formatter/base_spec.rb
@@ -3,6 +3,73 @@ require 'logstash-logger'
 describe LogStashLogger::Formatter::Base do
   include_context "formatter"
 
+  describe "#call" do
+    context "when event is not cancelled" do
+      it "returns a formatted message" do
+        expect(subject).to receive(:format_event).once.with(instance_of(LogStash::Event)).and_call_original
+        expect(subject.call(severity, time, progname, message)).to be_a(LogStash::Event)
+      end
+    end
+
+    context "when event is cancelled" do
+      before(:each) do
+        LogStashLogger.configure do |config|
+          config.customize_event(&:cancel)
+        end
+      end
+
+      it "returns `nil`" do
+        expect(subject).not_to receive(:format_event)
+        expect(subject.call(severity, time, progname, message)).to be_nil
+      end
+    end
+
+    context "when exception is raised" do
+      before do
+        allow(subject).to receive(:error_logger).and_return(Logger.new('/dev/null'))
+        allow(subject).to receive(:format_event).and_throw
+      end
+
+      it "logs an exception" do
+        expect(subject).to receive(:log_error)
+        subject.call(severity, time, progname, message)
+      end
+
+      it "returns a failed to format message" do
+        expect(subject.call(severity, time, progname, message)).to eq(LogStashLogger::Formatter::Base::FAILED_TO_FORMAT_MSG)
+      end
+    end
+  end
+
+  describe '#force_utf8_encoding' do
+    let(:message) { "foo x\xc3x".force_encoding('ASCII-8BIT').freeze }
+    let(:event) { LogStash::Event.new("message" => message) }
+
+    it 'returns the same event' do
+      expect(subject.send(:force_utf8_encoding, event)).to eq(event)
+    end
+
+    it 'forces the event message to UTF-8 encoding' do
+      subject.send(:force_utf8_encoding, event)
+      updated_event_data = event.instance_variable_get(:@data)
+      expect(updated_event_data['message'].encoding.name).to eq('UTF-8')
+    end
+
+    context 'frozen message string' do
+      let(:message) { "foo x\xc3x".dup.force_encoding('ASCII-8BIT').freeze }
+
+      it 'returns the same event' do
+        expect(subject.send(:force_utf8_encoding, event)).to eq(event)
+      end
+
+      it 'forces the event message to UTF-8 encoding' do
+        subject.send(:force_utf8_encoding, event)
+        updated_event_data = event.instance_variable_get(:@data)
+        expect(updated_event_data['message'].encoding.name).to eq('UTF-8')
+      end
+    end
+  end
+
   describe "#build_event" do
     let(:event) { formatted_message }
 
@@ -59,13 +126,44 @@ describe LogStashLogger::Formatter::Base do
       end
 
       it "adds host" do
-        expect(event['host']).to eq(hostname)
+        expect(event['host']['hostname']).to eq(hostname)
       end
     end
 
     describe "timestamp" do
       it "ensures time is in ISO8601 format" do
         expect(event.timestamp).to eq(time.iso8601(3))
+      end
+    end
+
+    describe "long messages" do
+
+      context "message field is present" do
+        let(:message) { long_message }
+
+        it "truncates long messages when max_message_size is set" do
+          LogStashLogger.configure do |config|
+            config.max_message_size = 2000
+          end
+
+          expect(event['message'].size).to eq(2000)
+        end
+      end
+
+      context "event without message field" do
+        let(:message) do
+          { 'test_field' => 'test', 'foo' => 'bar' }
+        end
+
+        it "still works" do
+          LogStashLogger.configure do |config|
+            config.max_message_size = 2000
+          end
+
+          expect(event['message']).to eq(nil)
+          expect(event['test_field']).to eq('test')
+          expect(event['foo']).to eq('bar')
+        end
       end
     end
   end