+++ spec/multi_logger_spec.rb
@@ -56,4 +56,20 @@ describe LogStashLogger::MultiLogger do
       logger.info 'test'
     end
   end
+
+  it "can call .reset on each logger" do
+    subject.loggers.each do |logger|
+      expect(logger).to receive(:reset).and_call_original
+    end
+
+    subject.reset
+  end
+
+  it "delegates #log to loggers" do
+    subject.loggers.each do |logger|
+      expect(logger).to receive(:add).with(::Logger::DEBUG, "test", nil)
+    end
+
+    subject.log(::Logger::DEBUG, "test")
+  end
 end